[sw/silicon_creator] Refactor bootstrap and reflash libraries
Now, bootstrap() in bootstrap.c checks whether bootstrap is requested,
but delegates the SPI loop to the common bootstrap library.

Issue #19151
