+++ sw/device/silicon_creator/lib/bootstrap.c
@@ -2,23 +2,19 @@
 // Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
 
+#include "sw/device/silicon_creator/lib/bootstrap.h"
+
 #include <stdalign.h>
 
 #include "sw/device/lib/base/abs_mmio.h"
 #include "sw/device/lib/base/bitfield.h"
 #include "sw/device/lib/base/hardened.h"
-#include "sw/device/silicon_creator/lib/base/chip.h"
 #include "sw/device/silicon_creator/lib/drivers/flash_ctrl.h"
-#include "sw/device/silicon_creator/lib/drivers/otp.h"
 #include "sw/device/silicon_creator/lib/drivers/rstmgr.h"
 #include "sw/device/silicon_creator/lib/drivers/spi_device.h"
 #include "sw/device/silicon_creator/lib/error.h"
-#include "sw/device/silicon_creator/rom/bootstrap.h"
 
 #include "flash_ctrl_regs.h"
-#include "gpio_regs.h"
-#include "hw/top_earlgrey/sw/autogen/top_earlgrey.h"
-#include "otp_ctrl_regs.h"
 
 enum {
   /*
@@ -337,33 +333,9 @@ static rom_error_t bootstrap_handle_program(bootstrap_state_t *state) {
   return error;
 }
 
-hardened_bool_t bootstrap_requested(void) {
-  uint32_t bootstrap_dis =
-      otp_read32(OTP_CTRL_PARAM_OWNER_SW_CFG_ROM_BOOTSTRAP_DIS_OFFSET);
-  if (launder32(bootstrap_dis) == kHardenedBoolTrue) {
-    return kHardenedBoolFalse;
-  }
-  HARDENED_CHECK_NE(bootstrap_dis, kHardenedBoolTrue);
-
-  // A single read is sufficient since we expect strong pull-ups on the strap
-  // pins.
-  uint32_t res = launder32(kHardenedBoolTrue) ^ SW_STRAP_BOOTSTRAP;
-  res ^=
-      abs_mmio_read32(TOP_EARLGREY_GPIO_BASE_ADDR + GPIO_DATA_IN_REG_OFFSET) &
-      SW_STRAP_MASK;
-  if (launder32(res) != kHardenedBoolTrue) {
-    return kHardenedBoolFalse;
-  }
-  HARDENED_CHECK_EQ(res, kHardenedBoolTrue);
-  return res;
-}
-
-rom_error_t bootstrap(void) {
-  hardened_bool_t requested = bootstrap_requested();
-  if (launder32(requested) != kHardenedBoolTrue) {
-    return kErrorBootstrapNotRequested;
-  }
-  HARDENED_CHECK_EQ(requested, kHardenedBoolTrue);
+rom_error_t enter_bootstrap(hardened_bool_t protect_rom_ext) {
+  // TODO(lowRISC/opentitan#19151): Implement `protect_rom_ext` behavior.
+  HARDENED_CHECK_EQ(protect_rom_ext, kHardenedBoolFalse);
 
   spi_device_init();
